Sat, Nov 30, 2024, 9:22 AM +03:00

+++ volume/volume.c
@@ -41,7 +41,7 @@ int main(int argc, char *argv[])
 
 
     // TODO: Read samples from input file and write updated data to output file
-    BYTE buffer;
+    int16_t buffer;
     while (fread(&buffer,1,2,input) !=0)
     {
         buffer *= factor;
